% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsegverse-package.R
\docType{package}
\name{ggsegverse-package}
\alias{ggsegverse}
\alias{ggsegverse-package}
\title{ggsegverse: Easily Install and Load the 'ggseg' Ecosystem}
\description{
The 'ggseg' ecosystem is a set of packages for brain region visualization using 'ggplot2'. This package makes it easy to install and load all core packages in a single call. Learn more about the 'ggseg' ecosystem at \url{https://ggseg.github.io/ggseg/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ggseg.github.io/ggsegverse/}
  \item \url{https://github.com/ggseg/ggsegverse}
  \item Report bugs at \url{https://github.com/ggseg/ggsegverse/issues}
}

}
\author{
\strong{Maintainer}: Athanasia Mo Mowinckel \email{a.m.mowinckel@psykologi.uio.no} (\href{https://orcid.org/0000-0002-5756-0223}{ORCID})

}
\keyword{internal}
