% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{ggsegverse_deps}
\alias{ggsegverse_deps}
\title{List ggsegverse package dependencies and versions}
\usage{
ggsegverse_deps()
}
\value{
A data frame with columns: \code{package}, \code{local}, \code{available},
\code{behind}.
}
\description{
List ggsegverse package dependencies and versions
}
