% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ggsegverse_packages}
\alias{ggsegverse_packages}
\title{List all ggsegverse packages}
\usage{
ggsegverse_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Whether to include ggsegverse itself in the list.}
}
\value{
A character vector of package names.
}
\description{
List all ggsegverse packages
}
